Change: tile change is now based on coordinates.
Before, the snake head entering a new tile was calculated by counting frames.
This was unreliable in case the snake was put in a misalligned position.
It was especially noticeable when respawning after a failstate.
Now, the tile change is calculated by checking if the x and y coords are divisible by the tile size.
Much more elegant and reliable approach.