Fix: dash function doesnt break the game
Hold down space to double your speed until you release the key.
This function broke the snake half the time because of the head's position not lining up.
The function now waits until the head is lined up before applying boost.
Added a disgustinly long if statement to do this.